Move NewDbusConn to testutils
Now that we are using it 3 times, factor it out.
