+++ internal/testutils/systembus.go
@@ -7,6 +7,10 @@ import (
 	"os/exec"
 	"path/filepath"
 	"sync"
+	"testing"
+
+	"github.com/godbus/dbus/v5"
+	"github.com/stretchr/testify/require"
 )
 
 var (
@@ -100,4 +104,23 @@ func StartLocalSystemBus() func() {
 		// Restore dbus system launcher
 		sdbus = sync.Once{}
 	}
+}
+
+// NewDbusConn returns a system dbus connection which will be tore down when tests ends
+func NewDbusConn(t *testing.T) *dbus.Conn {
+	t.Helper()
+
+	bus, err := dbus.SystemBusPrivate()
+	require.NoError(t, err, "Setup: can’t get a private system bus")
+
+	t.Cleanup(func() {
+		err = bus.Close()
+		require.NoError(t, err, "Teardown: can’t close system dbus connection")
+	})
+	err = bus.Auth(nil)
+	require.NoError(t, err, "Setup: can’t auth on private system bus")
+	err = bus.Hello()
+	require.NoError(t, err, "Setup: can’t send hello message on private system bus")
+
+	return bus
 }
